+++ kurtosis-lambda/impl/consts.ts
@@ -1,0 +1,3 @@
+export const TCP_PROTOCOL: string = "tcp";
+export const DOCKER_PORT_PROTOCOL_SEPARATOR: string = "/";
+export const EXEC_COMMAND_SUCCESS_EXIT_CODE: number = 0;